OLH-2581: Reinstate Welsh translations
Reinstate Welsh translations for the Rural Payment Wales service which
is obviously available in Welsh.

The service was accidentally flagged as not available in Welsh, and the
Welsh translations were removed. This reinstates the Welsh translations
for this service. Translations have been copied over from the time this
service was initially onboarded.