OLH-2385: Show 'Apply for teacher training' service in searchable list
This service onboarded at about the same time as we migrated to the
RP registry library and were working on the searchable list.

There was some confusion about Welsh translations at the time so it
looks like we postponed the decision to show this service in the
searchable list, and then it got dropped in the migration.